requirements.txt
# Core dependencies
python-telegram-bot==20.7
requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.0

# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Optional but recommended
aiohttp==3.9.1

# Redis for caching
redis==5.0.1

# Numpy for embeddings
numpy==1.26.3

# LangChain and FAISS for vector store
langchain==0.1.0
langchain-community==0.0.20
faiss-cpu==1.8.0
tiktoken==0.5.2

# Document loaders and parsers
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
